responding to PR coments
