+++ lib/stacks/iam.ts
@@ -50,8 +50,8 @@ export class IAMStack extends Stack {
     public mlSpacePermissionsBoundary?: IManagedPolicy;
     public emrServiceRoleName: string;
     public emrEC2RoleName: string;
-    public mlspaceEndpointConfigInstanceConstraintPolicy: IManagedPolicy;
-    public mlspaceJobInstanceConstraintPolicy: IManagedPolicy;
+    public mlspaceEndpointConfigInstanceConstraintPolicy?: IManagedPolicy;
+    public mlspaceJobInstanceConstraintPolicy?: IManagedPolicy;
 
     constructor (parent: App, name: string, props: IAMStackProp) {
         super(parent, name, {
@@ -434,36 +434,6 @@ export class IAMStack extends Stack {
             description: 'Enables general MLSpace actions in notebooks and across the entire application.'
         });
 
-        if (props.mlspaceConfig.ENDPOINT_CONFIG_INSTANCE_CONSTRAINT_POLICY_ARN) {
-            this.mlspaceEndpointConfigInstanceConstraintPolicy = ManagedPolicy.fromManagedPolicyArn(this, 'mlspace-endpoint-config-instance-constraint', props.mlspaceConfig.ENDPOINT_CONFIG_INSTANCE_CONSTRAINT_POLICY_ARN);
-        } else {
-            /*
-             * WARNING: @see instanceConstraintPolicyStatement
-             */
-            this.mlspaceEndpointConfigInstanceConstraintPolicy = new ManagedPolicy(this, 'mlspace-endpoint-config-instance-constraint', {
-                managedPolicyName: `${props.mlspaceConfig.IAM_RESOURCE_PREFIX}-endpoint-instance-constraint`,
-                statements: instanceConstraintPolicyStatement(this.partition, Aws.REGION, {CreateEndpointConfig: 'endpoint-config'})
-            });
-        }
-
-        if (props.mlspaceConfig.JOB_INSTANCE_CONSTRAINT_POLICY_ARN) {
-            this.mlspaceJobInstanceConstraintPolicy = ManagedPolicy.fromManagedPolicyArn(this, 'mlspace-job-instance-constraint', props.mlspaceConfig.JOB_INSTANCE_CONSTRAINT_POLICY_ARN);
-        } else {
-            /*
-             * WARNING: @see instanceConstraintPolicyStatement
-             */
-            this.mlspaceJobInstanceConstraintPolicy = new ManagedPolicy(this, 'mlspace-job-instance-constraint', {
-                managedPolicyName: `${props.mlspaceConfig.IAM_RESOURCE_PREFIX}-job-instance-constraint`,
-                statements: [
-                    instanceConstraintPolicyStatement(this.partition, Aws.REGION, {
-                        CreateHyperParameterTuningJob: 'hyper-parameter-tuning-job',
-                        CreateTrainingJob: 'training-job'
-                    })[0],
-                    instanceConstraintPolicyStatement(this.partition, Aws.REGION, {CreateTransformJob: 'transform-job'})[0]
-                ]
-            });
-        }
-
         if (props.mlspaceConfig.NOTEBOOK_ROLE_ARN) {
             this.mlSpaceNotebookRole = Role.fromRoleArn(
                 this,
@@ -471,6 +441,43 @@ export class IAMStack extends Stack {
                 props.mlspaceConfig.NOTEBOOK_ROLE_ARN
             );
         } else {
+            const managedPolicies: IManagedPolicy[] = [notebookPolicy];
+
+            if (props.mlspaceConfig.MANAGE_IAM_ROLES) {
+                if (props.mlspaceConfig.ENDPOINT_CONFIG_INSTANCE_CONSTRAINT_POLICY_ARN) {
+                    this.mlspaceEndpointConfigInstanceConstraintPolicy = ManagedPolicy.fromManagedPolicyArn(this, 'mlspace-endpoint-config-instance-constraint', props.mlspaceConfig.ENDPOINT_CONFIG_INSTANCE_CONSTRAINT_POLICY_ARN);
+                } else {
+                    /*
+                     * WARNING: @see instanceConstraintPolicyStatement
+                     */
+                    this.mlspaceEndpointConfigInstanceConstraintPolicy = new ManagedPolicy(this, 'mlspace-endpoint-config-instance-constraint', {
+                        managedPolicyName: `${props.mlspaceConfig.IAM_RESOURCE_PREFIX}-endpoint-instance-constraint`,
+                        statements: instanceConstraintPolicyStatement(this.partition, Aws.REGION, {CreateEndpointConfig: 'endpoint-config'})
+                    });
+                }
+        
+                if (props.mlspaceConfig.JOB_INSTANCE_CONSTRAINT_POLICY_ARN) {
+                    this.mlspaceJobInstanceConstraintPolicy = ManagedPolicy.fromManagedPolicyArn(this, 'mlspace-job-instance-constraint', props.mlspaceConfig.JOB_INSTANCE_CONSTRAINT_POLICY_ARN);
+                } else {
+                    /*
+                     * WARNING: @see instanceConstraintPolicyStatement
+                     */
+                    this.mlspaceJobInstanceConstraintPolicy = new ManagedPolicy(this, 'mlspace-job-instance-constraint', {
+                        managedPolicyName: `${props.mlspaceConfig.IAM_RESOURCE_PREFIX}-job-instance-constraint`,
+                        statements: [
+                            instanceConstraintPolicyStatement(this.partition, Aws.REGION, {
+                                CreateHyperParameterTuningJob: 'hyper-parameter-tuning-job',
+                                CreateTrainingJob: 'training-job'
+                            })[0],
+                            instanceConstraintPolicyStatement(this.partition, Aws.REGION, {CreateTransformJob: 'transform-job'})[0]
+                        ]
+                    });
+                }
+
+                managedPolicies.push(this.mlspaceEndpointConfigInstanceConstraintPolicy, this.mlspaceJobInstanceConstraintPolicy);
+            }
+            
+
             // If roles are managed by CDK, create the notebook role
             const mlSpaceNotebookRoleName = 'mlspace-notebook-role';
             // Translate Permissions Principles
@@ -484,11 +491,7 @@ export class IAMStack extends Stack {
             this.mlSpaceNotebookRole = new Role(this, 'mlspace-notebook-role', {
                 roleName: mlSpaceNotebookRoleName,
                 assumedBy: notebookPolicyAllowPrinciples,
-                managedPolicies: [
-                    notebookPolicy,
-                    this.mlspaceEndpointConfigInstanceConstraintPolicy,
-                    this.mlspaceJobInstanceConstraintPolicy
-                ],
+                managedPolicies,
                 description:
                     'Allows SageMaker Notebooks within ML Space to access necessary AWS services (S3, SQS, DynamoDB, ...)',
             });
